re-organized code and updated comments
