+++ src/scripts/posts/EditPost.js
@@ -1,19 +1,15 @@
 // Module goal: allow user to edit a saved post
 
-import { getSelectedPost, editPost } from "./PostProvider.js"
-import { checkDate } from "./PostBoxHTML.js"
+import { editPost } from "./PostProvider.js"
+import { EditPostForm } from "./EditPostHTML.js"
 
 const eventHub = document.querySelector(".container")
 
-// Stores original post so Cancel Button can re-render correct HTML
+// Stores original post so functions can access it. Allows Cancel Button to re-render correct HTML
 let originalPost = "";
 
 // Allows only one post to be editable at a time. Otherwise, originalPost can be overwritten
 let currentlyEditing = false;
-
-// I THINK WE CAN REMOVE THE CUSTOM EVENT
-// AND HAVE ALL THE SAVING OF INFORMATION HAPPEN
-// IN IMMEDIATE EDIT BTN PRESS
 
 // Custom event for when user clicks Edit Button:
     // saves original message, so user can cancel editing
@@ -46,7 +42,7 @@ eventHub.addEventListener("click", e => {
     const clicked = e.target.id
     if (clicked.startsWith("post__btnSave--")) {
         const [prefix, id] = clicked.split("--")
-        // Saves all edited data
+        // Save all edited data
         const editedPost = { 
             id: parseInt(document.querySelector(`#postEdit--postId--${id}`).value),
             post: document.querySelector(`#postEdit--text--${id}`).value,
@@ -57,7 +53,7 @@ eventHub.addEventListener("click", e => {
         editPost(editedPost, id)
         currentlyEditing = false;
     }
-    // Resets post to original state
+    // Reset post to original state
     if (clicked.startsWith("post__btnCancel--")) {
         e.target.parentElement.innerHTML = originalPost
         currentlyEditing = false;
@@ -67,26 +63,9 @@ eventHub.addEventListener("click", e => {
     if (clicked === "post__btnPost" && document.querySelector(".new__textarea").value !== "" || clicked.startsWith("post__btnDelete")) {
         currentlyEditing = false
     }
-    // Fixes bug when user clicks Post Button without entering text,
-    // currentlyEditing reverts to false without a re-render and multiple edit boxes can appear
+    // Fixes bug when user clicks Post Button without entering text.
+    // currentlyEditing reverts to false without a re-render, allowing multiple edit boxes to appear
     if (clicked === "post__btnPost" && document.querySelector(".new__textarea").value === "" && currentlyEditing === true) {
         currentlyEditing = true;
     }
 })
-
-// Generates HTML for post editing form
-const EditPostForm = postId => {
-    getSelectedPost(postId)
-        .then((response) => {
-            document.querySelector(`#postId--${postId}`).innerHTML = `
-            <input type="hidden" value="${response.id}" id="postEdit--postId--${postId}">
-            <input type="hidden" value="${response.userId}" id="postEdit--userId--${postId}">
-            <input type="hidden" value="${response.originalTimeStamp}" id="postEdit--originalTimeStamp--${postId}">
-            <span class="single__username">${response.user.username}</span>
-            ${checkDate(response)}
-            <textarea id="postEdit--text--${postId}">${response.post}</textarea>
-            <button id="post__btnSave--${postId}" type="button">Save Edit</button>
-            <button id="post__btnCancel--${postId}" type="button">Cancel Edit</button>
-            `
-        })
-}
